Start work on making the D3 Tools 64-bit safe.
* Fix RC Compile failure caused by forcing /MP flag

* Allow conan to be used to fulfill dependencies.  (Useful with MSVC on Windows)

* Significant codebase patching for 64-bit operation:

  * Fix the OnToolHitTest signature to match MFC

  * Change all the pointer storage from using Get/SetWindowLong to using Get/SetWindowLongPtr

  * Fixed intermediate casting for the above to prevent truncations.

* Adjust the id Dictionary to use int64_t for ints to prevent truncation of pointers or handles stowed

+++ neo/game/IK.cpp
@@ -443,7 +443,7 @@ bool idIK_Walk::Init( idEntity *self, const char *anim, const idVec3 &modelOffse
 		return false;
 	}
 
-	numLegs = Min( self->spawnArgs.GetInt( "ik_numLegs", "0" ), MAX_LEGS );
+	numLegs = Min( (int)self->spawnArgs.GetInt( "ik_numLegs", "0" ), MAX_LEGS );
 	if ( numLegs == 0 ) {
 		return true;
 	}
@@ -965,7 +965,7 @@ bool idIK_Reach::Init( idEntity *self, const char *anim, const idVec3 &modelOffs
 		return false;
 	}
 
-	numArms = Min( self->spawnArgs.GetInt( "ik_numArms", "0" ), MAX_ARMS );
+	numArms = Min( (int)self->spawnArgs.GetInt( "ik_numArms", "0" ), MAX_ARMS );
 	if ( numArms == 0 ) {
 		return true;
 	}
